refactor: input과 output 분리
