+++ src/main/java/lotto/io/output/ResultPrintHandler.java
@@ -1,4 +1,4 @@
-package lotto.io;
+package lotto.io.output;
 
 import lotto.util.ResultMessages;
